+++ HenkelEngine/scr/ECS/System/RenderSystem.cpp
@@ -33,7 +33,7 @@ void RenderSystem::Update()
 			renderComponent->SetQuadUVs(spriteComponent->tileSheet.GetSpriteRectAtIndex(spriteComponent->index));
 		}
 
-		glm::mat4 projection = m_engine->GetCurrentScene()->GetCamera()->CalculateProjection(m_engine->GetWindow()->GetWidth(), m_engine->GetWindow()->GetHeight());
+		glm::mat4 projection = m_engine->GetCurrentScene()->GetCamera()->CalculateProjection((float)m_engine->GetWindow()->GetWidth(), (float)m_engine->GetWindow()->GetHeight());
 		glm::mat4 view = m_engine->GetCurrentScene()->GetCamera()->GetViewMatrix();
 		glm::mat4 model = transformComponent->GetWorldMatrix();
 		materialComponent->Bind(model, view, projection);
